Refactor configuration and dependencies; remove unused data recorder
Problem
=======
Very slow and convoluted wrapper on top of a wrapper. Confusion if multiple objects could exisit and other problems.

Solution
========
Now, it does not depent on the API. THIS is the api itself, both the nodes are seperate and performs all the operations seperately.

Note
====
Lauch file added.
Warning: This is not yet tested.